moved confirm_add_song to different file
